Extract the degraded/park gates to a runtime mixin
Mechanical move: restores the 2,000-line file cap and the 100-line
__init__ cap after the park gate landed; signatures, attribute names,
and behavior unchanged.
